+++ app/javascripts/design-example.mjs
@@ -1,3 +1,4 @@
+import ClipboardJS from 'clipboard'
 import { iframeResize } from 'iframe-resizer'
 
 class DesignExample {
@@ -13,8 +14,8 @@ class DesignExample {
 
     this.tabs = this.node.querySelectorAll(`.${this.tabClass}`)
     this.examples = this.node.querySelectorAll('.code-snippet__preformatted')
-    this.closeButtons = this.node.querySelectorAll('.app-link--close')
-    this.copyButtons = this.node.querySelectorAll('.app-link--copy')
+    this.closeButtons = this.node.querySelectorAll('.app-button--close')
+    this.copyButtons = this.node.querySelectorAll('.app-button--copy')
     this.iframe = this.node.querySelector('iframe')
 
     this.bindEvents()
@@ -27,13 +28,14 @@ class DesignExample {
       tab.addEventListener('click', (e) => this.handleTabClick(e))
     )
 
-    this.closeButtons.forEach((closeButton) =>
+    this.closeButtons.forEach((closeButton) => {
       closeButton.addEventListener('click', (e) => this.handleCloseClick(e))
-    )
+      closeButton.removeAttribute('hidden')
+    })
 
-    this.copyButtons.forEach((copyButton) =>
-      copyButton.addEventListener('click', (e) => this.handleCopyClick(e))
-    )
+    if (ClipboardJS.isSupported()) {
+      this.copyButtons.forEach((copyButton) => this.initCopyClick(copyButton))
+    }
   }
 
   handleTabClick(e) {
@@ -65,20 +67,24 @@ class DesignExample {
     })
   }
 
-  // Follows approach from https://hackernoon.com/copying-text-to-clipboard-with-javascript-df4d4988697f
-  handleCopyClick(e) {
-    e.preventDefault()
-    const el = document.createElement('textarea')
-    const str = this.node.querySelector(
-      '.code-snippet__preformatted:not(.js-hidden) code'
-    )
-    el.value = str.innerText
-    document.body.appendChild(el)
-    el.select()
-    document.execCommand('copy')
-    document.body.removeChild(el)
-    e.target.innerText = 'Copied'
-    setTimeout(() => (e.target.innerText = 'Copy code'), 2500)
+  initCopyClick(copyButton) {
+    const clipboard = new ClipboardJS(copyButton, {
+      target: () => copyButton.parentElement.querySelector('pre')
+    })
+
+    // Update button on success
+    clipboard.on('success', (event) => {
+      copyButton.innerText = 'Code copied'
+      event.clearSelection()
+
+      // Reset button after delay
+      setTimeout(() => {
+        copyButton.innerText = 'Copy code'
+      }, 2500)
+    })
+
+    // Reveal button
+    copyButton.removeAttribute('hidden')
   }
 
   showEl(el) {